requirements.txt
django
django-webpack-loader
psycopg2
twarc
rq
django-allauth
